# openapi/v1.yaml
...
components:
  schemas:
...
    CurrentUserOrganization:
      description: Organization summary embedded in the current-user response.
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
      required:
      - id
      - name
      title: CurrentUserOrganization
      type: object
...
    LoginUser:
      description: User summary embedded in a successful login response.
      properties:
        email:
          title: Email
          type: string
        id:
          title: Id
          type: integer
        organization_id:
          title: Organization Id
          type: integer
        role:
          title: Role
          type: string
        username:
          title: Username
          type: string
      required:
      - id
      - username
      - email
      - role
      - organization_id
      title: LoginUser
      type: object
...
    VerifyUser:
      description: User summary embedded in the auth-verify response.
      properties:
        id:
          title: Id
          type: integer
        organization_id:
          title: Organization Id
          type: integer
        role:
          title: Role
          type: string
        username:
          title: Username
          type: string
      required:
      - id
      - username
      - role
      - organization_id
      title: VerifyUser
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: WaddleAI Management API
  version: '1'
openapi: 3.1.0
paths:
  /api/v1/auth/change-password:
    post:
      description: ''
      operationId: post_change_password
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              description: Request body for POST /api/v1/auth/change-password.
              properties:
                current_password:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: Current Password
                new_password:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: New Password
              title: ChangePasswordRequest
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'Generic `{"message": str}` envelope used by several
                  auth endpoints.'
                properties:
                  message:
                    title: Message
                    type: string
                required:
                - message
                title: MessageResponse
                type: object
          description: 'Generic `{"message": str}` envelope used by several auth endpoints.'
      security:
      - bearerAuth: []
      summary: Change user password
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      description: ''
      operationId: post_login
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              description: Credentials for POST /api/v1/auth/login.
              properties:
                password:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: Password
                username:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: Username
              title: LoginRequest
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for a successful login.
                properties:
                  access_token:
                    title: Access Token
                    type: string
                  expires_in:
                    title: Expires In
                    type: integer
                  token_type:
                    title: Token Type
                    type: string
                  user:
                    $ref: '#/components/schemas/LoginUser'
                required:
                - access_token
                - token_type
                - expires_in
                - user
                title: LoginResponse
                type: object
          description: Response body for a successful login.
      summary: User login endpoint
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: ''
      operationId: post_logout
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'Generic `{"message": str}` envelope used by several
                  auth endpoints.'
                properties:
                  message:
                    title: Message
                    type: string
                required:
                - message
                title: MessageResponse
                type: object
          description: 'Generic `{"message": str}` envelope used by several auth endpoints.'
      security:
      - bearerAuth: []
      summary: User logout endpoint
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: ''
      operationId: get_get_current_user
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for GET /api/v1/auth/me.
                properties:
                  created_at:
                    anyOf:
                    - type: string
                    - type: 'null'
                    title: Created At
                  email:
                    title: Email
                    type: string
                  enabled:
                    title: Enabled
                    type: boolean
                  id:
                    title: Id
                    type: integer
                  last_login_at:
                    anyOf:
                    - type: string
                    - type: 'null'
                    title: Last Login At
                  organization:
                    anyOf:
                    - $ref: '#/components/schemas/CurrentUserOrganization'
                    - type: 'null'
                  role:
                    title: Role
                    type: string
                  token_quota_daily:
                    anyOf:
                    - type: integer
                    - type: 'null'
                    title: Token Quota Daily
                  token_quota_monthly:
                    anyOf:
                    - type: integer
                    - type: 'null'
                    title: Token Quota Monthly
                  username:
                    title: Username
                    type: string
                required:
                - id
                - username
                - email
                - role
                - organization
                - token_quota_daily
                - token_quota_monthly
                - enabled
                - created_at
                - last_login_at
                title: CurrentUserResponse
                type: object
          description: Response body for GET /api/v1/auth/me.
      security:
      - bearerAuth: []
      summary: Get current user info
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      description: ''
      operationId: post_refresh_token
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for a successful token refresh.
                properties:
                  access_token:
                    title: Access Token
                    type: string
                  expires_in:
                    title: Expires In
                    type: integer
                  token_type:
                    title: Token Type
                    type: string
                required:
                - access_token
                - token_type
                - expires_in
                title: RefreshTokenResponse
                type: object
          description: Response body for a successful token refresh.
      security:
      - bearerAuth: []
      summary: Refresh JWT token
      tags:
      - Auth
  /api/v1/auth/verify:
    get:
      description: ''
      operationId: get_verify_auth
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for GET /api/v1/auth/verify.
                properties:
                  user:
                    $ref: '#/components/schemas/VerifyUser'
                required:
                - user
                title: VerifyResponse
                type: object
          description: Response body for GET /api/v1/auth/verify.
      security:
      - bearerAuth: []
      summary: Verify the caller's bearer token and echo back its identity claims.
      tags:
      - Auth
...